pyproject.toml
[tool.poetry]
name = "wallet_management_tools"
version = "0.1.0"
description = "Python 项目"
authors = ["Oxmoei <https://t.me/cryptostar210>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
alive-progress = "3.1.4"
loguru = "0.6.0"
requests = "2.28.2"
termcolor = "2.3.0"
tls-client = "0.2.1"
tabulate = "0.9.0"
setuptools = "65.0.0"
python-dotenv = "1.0.1"
web3 = "7.12.0"
httpx = "0.28.1"
solana = "0.36.6"
colorama = "^0.4.6"
solders = ">=0.23,<0.27"
anchorpy = "^0.21.0"
PyYAML = "^6.0"
websockets = "^12.0"
click = "^8.1.7"
rich = "^13.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
